requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
pandas==2.1.4
openpyxl==3.1.2
python-multipart==0.0.6
duckdb==0.10.0
pyarrow==15.0.0
great-expectations==0.18.8
numpy==1.26.3
scikit-learn==1.4.0
